tools/midi2click/requirements.txt
mido>=1.3.0
python-rtmidi>=1.5.0
pyobjc-framework-Quartz>=9.0
